requirements.txt
Flask==2.3.2
spacy==3.7.2
scikit-learn==1.3.2
pdfminer.six==20221105
python-docx==0.8.11
pandas==2.1.4
fpdf2==2.7.4
python-dateutil==2.8.2
joblib==1.3.2
nltk==3.8.1
gunicorn==21.2.0
https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.7.0/en_core_web_lg-3.7.0-py3-none-any.whl
